Moves symbol table map and special into new module
