--- yoyodyne/special.py
@@ -1,10 +1,0 @@
-"""Reserved symbols."""
-
-# These are all reserved for internal use.
-
-PAD = "<P>"
-START = "<S>"
-END = "<E>"
-UNK = "<UNK>"
-
-SPECIAL = [UNK, PAD, START, END]
